Fixed bug on android when bounds are requested and _map is null
If bounds are requested and onMapReady is not yet called, _map is null.
Previously in such situation LatLngBounds are creater using builder but without
any LatLng item, so builder would throw an exception saying "Can not create LatLngBounds with 0 items". Now initial center LatLng is repeated twice, since two items are required. This is still better then throwing exception that user can not understand or control.