fixes crash on Android re-enter app
https://github.com/mapbox/react-native-mapbox-gl/issues/999

The MapView has a locationPlugin that is registered with the locationEngine, and it never got removed when the view was torn down. This will tell the LocationPlugin to stop when the view is detached. 

However, the Lost Engine is still active in the background. This patch simply lets you press back to close the app, and then re-enter without a crash